Add files via upload
This MASTER SCRIPT has been used to produce the analysis presented in "Barrington et al. (2022) Wind speed as a dominant source of periodicities in reported emission rates of volcanic SO2" and has been provided to in the interest of research integrity. This version of the code is not thoroughly commented nor is it free of redundancies. Figures produced using this code have been modified and formatted independently but the data unchanged. Users wishing to apply similar analysis to their own datasets, are directed to use the user-friendly snippets of code that are provided. 

This ‘MasterScript’ calls on the functions contained in‘MasterScript_functions’.